requirements.txt
brotli==1.1.0
ConfigParser==6.0.0
cryptography==41.0.7
Cython==3.0.6
dl==0.1.0
docutils==0.20.1
Flask==3.0.0
HTMLParser==0.0.2
importlib_metadata==7.0.0
jnius==1.1.0
keyring==24.3.0
lockfile==0.12.2
lxml==4.9.3
numpy==1.26.2
ordereddict==1.1
protobuf==4.25.1
psycopg2_binary==2.9.9
pyOpenSSL==23.3.0
python-dotenv==1.0.0
thread==0.1.2
toml==0.10.2
tornado==6.4
typing_extensions==4.8.0
xmlrpclib==1.0.1
zipp==3.17.0
